+++ tests/oauth2/rfc6749/endpoints/test_revocation_endpoint.py
@@ -88,19 +88,6 @@ class RevocationEndpointTest(TestCase):
         self.assertEqual(loads(b)['error'], 'invalid_client')
         self.assertEqual(s, 401)
 
-    def test_revoke_with_callback(self):
-        endpoint = RevocationEndpoint(self.validator, enable_jsonp=True)
-        callback = 'package.hello_world'
-        for token_type in ('access_token', 'refresh_token', 'invalid'):
-            body = urlencode([('token', 'foo'),
-                              ('token_type_hint', token_type),
-                              ('callback', callback)])
-            h, b, s = endpoint.create_revocation_response(self.uri,
-                    headers=self.headers, body=body)
-            self.assertEqual(h, {})
-            self.assertEqual(b, callback + '();')
-            self.assertEqual(s, 200)
-
     def test_revoke_unsupported_token(self):
         endpoint = RevocationEndpoint(self.validator,
                                       supported_token_types=['access_token'])