+++ CHANGELOG.rst
@@ -3,6 +3,9 @@ Changelog
 
 WIP
 ---
+OAuth2.0 Provider:
+* **Breaking**: Removed JSONP support from token revocation endpoint. JSONP has been superseded by CORS for cross-origin requests. The ``enable_jsonp`` parameter has been removed from ``RevocationEndpoint`` and the ``callback`` parameter has been removed from ``prepare_token_revocation_request``.
+
 Misc:
 * #930: Add devcontainer, Add Python3.14, Python3.14t.